increse legibility

+++ script/generator.rb
@@ -5,51 +5,57 @@ require './tool-http'
 
 http_tool = ToolHttp.new
 
-puts 'Enter number of files: [10]'
-files = gets.chomp
-if files.to_s.empty?
-	files = 10
-else
-	files = files.to_i
-end
+args = Hash[ ARGV.join(' ').scan(/--?([^=\s]+)(?:=(\S+))?/) ]
 
-puts 'Enter root of files: [/Users/power/Tmp/]'
-root = gets.chomp
-if root.to_s.empty?
-	root = "/Users/power/Tmp/"
-end
+if args.key?('new')
 
-# delete all files generated in the previous execution
-Dir.foreach(root) do |file|
-	if ((file.to_s != ".") and (file.to_s != ".."))
-		File.delete("#{root}/#{file}")
+	puts 'Enter number of files: [10]'
+	files = gets.chomp
+	if files.to_s.empty?
+		files = 10
+	else
+		files = files.to_i
 	end
-end
 
-for number in 1..files
-	file_name = SecureRandom.hex
-
-	File.open(root + file_name, 'w') do |f|
-		f.print file_name
+	puts 'Enter root of files: [/Users/power/Tmp/]'
+	root = gets.chomp
+	if root.to_s.empty?
+		root = "/Users/power/Tmp/"
 	end
-end
 
-puts 'Enter path where to write configuration file: [/Users/power/Dev/github/kilebeat/confs/]'
-conf_file = gets.chomp
-if conf_file.to_s.empty?
-	conf_file = "/Users/power/Dev/github/kilebeat/confs/"
-end
-conf_file += SecureRandom.hex
-puts 'New configuration file will be write in ' + conf_file
+	# delete all files generated in the previous execution
+	Dir.foreach(root) do |file|
+		if ((file.to_s != ".") and (file.to_s != ".."))
+			File.delete("#{root}/#{file}")
+		end
+	end
 
-File.open(conf_file, 'w') do |f|
-	@elements = Dir.foreach(root).reject {|file| file.to_s == "." or file.to_s == ".."}
-	f.print ERB.new(File.open("template.erb", 'r').read).result(binding)
-end
+	for number in 1..files
+		file_name = SecureRandom.hex
 
-#File.delete(File.readlink("/Users/power/Dev/github/kilebeat/kilebeat.conf"))
-File.unlink("/Users/power/Dev/github/kilebeat/kilebeat.conf")
-File.symlink(conf_file, "/Users/power/Dev/github/kilebeat/kilebeat.conf")
+		File.open(root + file_name, 'w') do |f|
+			f.print file_name
+		end
+	end
+
+	puts 'Enter path where to write configuration file: [/Users/power/Dev/github/kilebeat/confs/]'
+	conf_file = gets.chomp
+	if conf_file.to_s.empty?
+		conf_file = "/Users/power/Dev/github/kilebeat/confs/"
+	end
+	conf_file += SecureRandom.hex
+	puts 'New configuration file will be write in ' + conf_file
+
+	File.open(conf_file, 'w') do |f|
+		@elements = Dir.foreach(root).reject {|file| file.to_s == "." or file.to_s == ".."}
+		f.print ERB.new(File.open("template.erb", 'r').read).result(binding)
+	end
+
+	#File.delete(File.readlink("/Users/power/Dev/github/kilebeat/kilebeat.conf"))
+	File.unlink("/Users/power/Dev/github/kilebeat/kilebeat.conf")
+	File.symlink(conf_file, "/Users/power/Dev/github/kilebeat/kilebeat.conf")
+
+end
 
 #REPL start Here
 loop do
